hotfix: lifecicle for quote request

+++ app/Http/Controllers/QuoteController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\StoreQuoteRequest;
 use App\Http\Requests\UpdateQuoteRequest;
 use App\Models\Quote;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Facades\Auth;
 
 class QuoteController extends Controller
 {
@@ -13,10 +14,21 @@ class QuoteController extends Controller
      * Display a listing of the resource.
      */
     public function index()
-    {
-        $Quotes = Quote::all();
-        return view('quote.index')->with('quotes', $Quotes);
+{
+    if (Auth::user()->can('viewAny', Quote::class)) {
+        if (Auth::user()->hasRole('admin')) {
+            // Si el usuario tiene el rol de administrador, obtén todas las citas
+            $quotes = Quote::all();
+        } else {
+            // Si el usuario no es administrador, obtén solo las citas asociadas a su correo electrónico
+            $quotes = Quote::where('email', Auth::user()->email)->get();
+        }
+    
+        return view('quote.index', compact('quotes'));
+    } else {
+        abort(403, 'Unauthorized action.');
     }
+}
 
 
     public function create()
@@ -32,7 +44,7 @@ class QuoteController extends Controller
         $Quote->phone = $request->input('phone');
 
         if ($request->has('phone')) {
-            $Quote->email = $request->input('email');
+            $Quote->phone = $request->input('phone');
         }
 
         $Quote->description = $request->input('description');
@@ -56,23 +68,30 @@ class QuoteController extends Controller
 
         $Quote->save();
 
-        return redirect()->intended(route('quotes.index'))->with('success', 'Your quote request has been successfully created. We will contact you as soon as possible.');
+        if (Auth::check()) {
+            return redirect()->route('quotes.index')->with('success', 'Your quote request was submitted successfully. You can view your requests and their status here.');
+        } else {
+            return redirect()->route('login')->with('success', 'Your quote request was submitted successfully. To track it, please log in by creating a new account or access it through your Google account for faster processing.');
+        }
     }
 
 
     public function show(Quote $quote)
     {
+        $this->authorize('view', $quote);
         return view('quote.show')->with('quote', $quote);
     }
 
     public function edit(Quote $quote)
     {
+        $this->authorize('update', $quote);
         return view('quote.edit')->with('quote', $quote);
     }
 
 
     public function update(UpdateQuoteRequest $request, Quote $Quote)
     {
+        $this->authorize('update', $Quote);
         if ($request->hasFile('image')) {
             $existingMedia = $Quote->getFirstMedia('quote_images');
         
@@ -119,6 +138,7 @@ class QuoteController extends Controller
 
     public function destroy(Quote $Quote)
     {
+        $this->authorize('delete', $Quote);
         if ($Quote->hasMedia('quote_images')) {
             $Quote->getMedia('quote_images')->each(function ($media) {
                 Storage::delete($media->getPath());
